remove unused dithering utils
