+++ rubin_scheduler/scheduler/utils/__init__.py
@@ -1,5 +1,4 @@
 from .comcam_tessellate import *
-from .dithering import *
 from .footprints import *
 from .sky_area import *
 from .tsp import *
